data: Add "Lollipop" to desktop file keywords
As an en_GB speaker, I abbreviate the word to "lolly", but in full write
"lollipop".  As a result, when I use desktop search to find this app, I
often type "lolli" and get no results.

Wiktionary agrees with my spelling choices: it uses [lollipop][0] is the
canonical spelling of this delicious sweet, and describes [lollypop][1]
as an alternative spelling.

Add the spelling with an 'i' to the keywords in the desktop file, so
that "lollipop" will find the app in a desktop search. (This is one of a
[half-dozen fields][2] that are used for search.)

[0]: https://en.wiktionary.org/wiki/lollipop
[1]: https://en.wiktionary.org/wiki/lollypop
[2]: https://gitlab.gnome.org/GNOME/glib/blob/2.66.1/gio/gdesktopappinfo.c#L396-401